@@ -10,7 +10,7 @@
   "dependencies": {
     "axios": "^1.3.0",
     "core-js": "^3.8.3",
-    "vue": "^2.6.14"
+    "vue": "^3.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
@@ -18,9 +18,9 @@
     "@vue/cli-plugin-babel": "~5.0.0",
     "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-service": "~5.0.0",
+    "@vue/compiler-sfc": "^3.1.0",
     "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "vue-template-compiler": "^2.6.14"
+    "eslint-plugin-vue": "^8.0.3"
   },
   "eslintConfig": {
     "root": true,
