@@ -1,9 +1,4 @@
-import Vue from 'vue';
-import App from './App.vue';
+import { createApp } from 'vue'
+import App from './App.vue'
 
-//turn it later to false
-Vue.config.productionTip = true
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).mount('#app')
